removed todolist_id in TaskCreateDTO and made it a parameter in the endpoints and adjusted functions of endpoints to work in accordance to the change.Edited the events to return out of the function when certain conditions aren't met to aid efficiency

+++ routers/task_routers.py
@@ -30,32 +30,43 @@ def after_insert_todo(mapper, connection, target):
         todolist = session.get(TodoList, target.todolist_id)  
         if todolist:
             todolist.task_number += 1
-            session.commit() 
+            todolist.is_completed = False
+        session.commit() 
 
 @event.listens_for(Todo,"after_update")
 def after_task_completed(mapper, connection, target):
     if target.is_completed is True:
         with Session(bind=connection) as session:
             todolist = session.get(TodoList, target.todolist_id)  
-            if todolist:
-                todolist.completed_tasks += 1
-                if todolist.completed_tasks == todolist.task_number:
-                    todolist.is_completed = True
+            if not todolist:
+                return
+            todolist.completed_tasks += 1
+            session.commit()
+            if todolist.completed_tasks == 0 and todolist.task_number == 0:
+                todolist.is_completed = False
+                session.commit()
+                return
+            if todolist.completed_tasks == todolist.task_number:
+                todolist.is_completed = True
                 session.commit() 
 
 @event.listens_for(Todo,"before_update")
 def after_task_undone(mapper, connection, target):
     history = get_history(target, "is_completed")
-    if history.has_changes:
-        if history.deleted:
-            if history.deleted[-1] == True:
-                if target.is_completed is False:
-                    with Session(bind=connection) as session:
-                        todolist = session.get(TodoList, target.todolist_id)
-                        todolist.is_completed = False
-                        if todolist:
-                            todolist.completed_tasks -= 1
-                            session.commit() 
+    if  not history.has_changes:
+        return
+    if history.deleted:
+        if history.deleted[-1] == False:
+            return
+        if target.is_completed is True:
+            return
+        with Session(bind=connection) as session:
+            todolist = session.get(TodoList, target.todolist_id)
+            todolist.is_completed = False
+            if todolist:
+                todolist.completed_tasks -= 1
+            session.commit() 
+
 
 
 
@@ -67,6 +78,13 @@ def deleted_todo(mapper,connection,target):
         todolist.task_number -= 1
         if todolist.is_completed == True:
             todolist.completed_tasks -= 1
+            session.commit()
+        if todolist.completed_tasks == 0 and todolist.task_number == 0:
+            todolist.is_completed = False
+            session.commit()
+            return
+        if todolist.completed_tasks == todolist.task_number:
+            todolist.is_completed = True
         session.commit()
 
 
@@ -79,9 +97,9 @@ def get_session():
 
 
 @router1.post("",response_model=Todo, status_code = 201 )
-def create_task(task: TaskCreateDTO,response: Response, session: Session = Depends(get_session)):
-    db_item = Todo(**task.model_dump())  # created_at and updated_at will be set automatically by the event
-    todolist = session.get(TodoList, db_item.todolist_id)
+def create_task(task: TaskCreateDTO,todolist_id: int,response: Response, session: Session = Depends(get_session)):
+    db_item = Todo(**task.model_dump(), todolist_id=todolist_id)
+    todolist = session.get(TodoList, todolist_id)
     if todolist:   
         session.add(db_item)
         session.commit()
@@ -98,24 +116,33 @@ def read_tasks(session: Session = Depends(get_session)):
     tasks = session.exec(select(Todo)).all()  # Query the database for all Task records
     return tasks
 
+@router1.get("/todolist",response_model = List[Todo])
+def read_Todolist_tasks(todolist_id: int, session: Session = Depends(get_session)):
+    statement = select(Todo).where(Todo.todolist_id == todolist_id)      
+    tasks = session.exec(statement).all()
+    if not tasks:
+        raise HTTPException(status_code=404, detail="Task Not Found!")
+       
+    return tasks
 
 
-@router1.get("/{task_id}/", response_model=Todo)
-def read_task(task_id: int, session: Session = Depends(get_session)):
+@router1.get("/{task_id}", response_model=Todo)
+def read_task(task_id: int, todolist_id: int, session: Session = Depends(get_session)):
     task = session.get(Todo, task_id)
     if not task:
         raise HTTPException(status_code=404, detail="Task Not Found!")
+    if task.todolist_id != todolist_id:
+        raise HTTPException(status_code=404, detail="TodoList Not Found!") 
+    
     return task
 
 @router1.put("/{task_id}", response_model=Todo)
-def update_task(task_id: int, task: TaskCreateDTO, session: Session = Depends(get_session)):
+def update_task(task_id: int,todolist_id: int, task: TaskCreateDTO, session: Session = Depends(get_session)):
     task_item = session.get(Todo, task_id)
-    todolist = session.get(TodoList, task_item.todolist_id)
     if not task_item:
         raise HTTPException(status_code=404, detail="Task Not Found!")
-    if not todolist:
+    if task_item.todolist_id != todolist_id:
         raise HTTPException(status_code=404, detail="TodoList Not Found!")
-
     if task_item.is_completed is True:
         raise HTTPException(status_code=404,detail="Task is Completed")
     task_data = task.model_dump(exclude_unset=False)
@@ -127,10 +154,13 @@ def update_task(task_id: int, task: TaskCreateDTO, session: Session = Depends(ge
     return task_item
 
 @router1.patch("/{task_id}/complete", response_model=Todo)
-def complete_task(task_id: int, session: Session = Depends(get_session)):
+def complete_task(task_id: int,todolist_id:int, session: Session = Depends(get_session)):
     task_item = session.get(Todo, task_id)
     if not task_item:
         raise HTTPException(status_code=404, detail="Task Not Found!")
+    if task_item.todolist_id != todolist_id:
+        raise HTTPException(status_code=404, detail="TodoList Not Found!")
+    
     if task_item.is_completed:
         raise HTTPException(status_code=401, detail="Already Completed")
     task_item.is_completed = True
@@ -141,10 +171,14 @@ def complete_task(task_id: int, session: Session = Depends(get_session)):
     return task_item
 
 @router1.patch("/{task_id}/undo", response_model=Todo)
-def undo_task(task_id: int, session: Session = Depends(get_session)):
+def undo_task(task_id: int,todolist_id: int, session: Session = Depends(get_session)):
     task_item = session.get(Todo, task_id)
     if not task_item:
         raise HTTPException(status_code=404, detail="Task Not Found!")
+    
+    if task_item.todolist_id != todolist_id:
+        raise HTTPException(status_code=404, detail="TodoList Not Found!")
+     
     if not task_item.is_completed:
         raise HTTPException(status_code=401, detail="Cannot Undo")
     task_item.is_completed = False
@@ -156,10 +190,12 @@ def undo_task(task_id: int, session: Session = Depends(get_session)):
 
 
 @router1.delete("/{task_id}", status_code = 204)
-def delete_task(task_id: int, response: Response,session: Session = Depends(get_session)):
+def delete_task(task_id: int, todolist_id: int,response: Response,session: Session = Depends(get_session)):
     task = session.get(Todo, task_id)
     if not task:
         raise HTTPException(status_code=404, detail="Task hasn't been created")
+    if task.todolist_id != todolist_id:
+        raise HTTPException(status_code=404, detail="TodoList Not Found!")
     session.delete(task)
     session.commit()
     response.status_code = status.HTTP_204_NO_CONTENT
